src/python-consumer/requirements.txt
kafka-python>=2.0.1,<2.1
lxml>=4.5.2,<4.6
boto3>=1.14.54,<1.15
